Extract getCardImageUrl and remove duplicate hook
