+++ src/components/cardbuilder/Card/useCard.ts
@@ -2,6 +2,7 @@ import classNames from 'classnames';
 
 import layoutManager from 'components/layoutManager';
 import { ItemAction } from 'constants/itemAction';
+import { useApi } from 'hooks/useApi';
 import { ItemKind } from 'types/base/models/item-kind';
 import { ItemMediaKind } from 'types/base/models/item-media-kind';
 import type { ItemDto } from 'types/base/models/item-dto';
@@ -9,11 +10,11 @@ import type { CardOptions } from 'types/cardOptions';
 import { CardShape } from 'utils/card';
 import { getDataAttributes } from 'utils/items';
 
-import useCardImageUrl from './useCardImageUrl';
 import {
     resolveAction,
     resolveMixedShapeByAspectRatio
 } from '../cardBuilderUtils';
+import { getCardImageUrl } from '../utils/url';
 
 interface UseCardProps {
     item: ItemDto;
@@ -21,6 +22,7 @@ interface UseCardProps {
 }
 
 function useCard({ item, cardOptions }: UseCardProps) {
+    const { api } = useApi();
     const action = resolveAction({
         defaultAction: cardOptions.action ?? ItemAction.Link,
         isFolder: item.IsFolder ?? false,
@@ -33,15 +35,18 @@ function useCard({ item, cardOptions }: UseCardProps) {
         shape = resolveMixedShapeByAspectRatio(item.PrimaryImageAspectRatio);
     }
 
-    const imgInfo = useCardImageUrl({
+    const {
+        imgUrl,
+        blurhash,
+        forceName,
+        coverImage
+    } = getCardImageUrl({
+        api,
         item: item.ProgramInfo ?? item,
-        cardOptions,
+        options: cardOptions,
         shape
     });
-    const imgUrl = imgInfo.imgUrl;
-    const blurhash = imgInfo.blurhash ?? undefined;
-    const forceName = imgInfo.forceName;
-    const coveredImage = cardOptions.coverImage ?? imgInfo.coverImage;
+    const coveredImage = cardOptions.coverImage ?? coverImage;
     const overlayText = cardOptions.overlayText;
 
     const nameWithPrefix = item.SortName ?? item.Name ?? '';
@@ -112,7 +117,7 @@ function useCard({ item, cardOptions }: UseCardProps) {
         className: cardBoxClass,
         shape,
         imgUrl,
-        blurhash,
+        blurhash: blurhash ?? undefined,
         forceName,
         coveredImage,
         overlayText
